extract send message functionality to separate viewmodel
