+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/ConversationScreen.kt
@@ -116,6 +116,7 @@ import com.wire.android.ui.home.conversations.banner.ConversationBanner
 import com.wire.android.ui.home.conversations.banner.ConversationBannerViewModel
 import com.wire.android.ui.home.conversations.call.ConversationCallViewModel
 import com.wire.android.ui.home.conversations.call.ConversationCallViewState
+import com.wire.android.ui.home.conversations.composer.MessageComposerViewModel
 import com.wire.android.ui.home.conversations.delete.DeleteMessageDialog
 import com.wire.android.ui.home.conversations.details.GroupConversationDetailsNavBackArgs
 import com.wire.android.ui.home.conversations.edit.EditMessageMenuItems
@@ -130,6 +131,7 @@ import com.wire.android.ui.home.conversations.model.ExpirationStatus
 import com.wire.android.ui.home.conversations.model.UIMessage
 import com.wire.android.ui.home.conversations.selfdeletion.SelfDeletionMapper.toSelfDeletionDuration
 import com.wire.android.ui.home.conversations.selfdeletion.SelfDeletionMenuItems
+import com.wire.android.ui.home.conversations.sendmessage.SendMessageViewModel
 import com.wire.android.ui.home.gallery.MediaGalleryActionType
 import com.wire.android.ui.home.gallery.MediaGalleryNavBackArgs
 import com.wire.android.ui.home.messagecomposer.MessageComposer
@@ -188,6 +190,7 @@ fun ConversationScreen(
     conversationCallViewModel: ConversationCallViewModel = hiltViewModel(),
     conversationMessagesViewModel: ConversationMessagesViewModel = hiltViewModel(),
     messageComposerViewModel: MessageComposerViewModel = hiltViewModel(),
+    sendMessageViewModel: SendMessageViewModel = hiltViewModel(),
     conversationMigrationViewModel: ConversationMigrationViewModel = hiltViewModel(),
     messageDraftViewModel: MessageDraftViewModel = hiltViewModel(),
     groupDetailsScreenResultRecipient: ResultRecipient<GroupConversationDetailsScreenDestination, GroupConversationDetailsNavBackArgs>,
@@ -335,7 +338,7 @@ fun ConversationScreen(
                 NavigationCommand(MessageDetailsScreenDestination(conversationInfoViewModel.conversationId, messageId, isSelfMessage))
             )
         },
-        onSendMessage = messageComposerViewModel::trySendMessage,
+        onSendMessage = sendMessageViewModel::trySendMessage,
         onDeleteMessage = conversationMessagesViewModel::showDeleteMessageDialog,
         onAssetItemClicked = conversationMessagesViewModel::downloadOrFetchAssetAndShowDialog,
         onImageFullScreenMode = { message, isSelfMessage ->
@@ -389,14 +392,14 @@ fun ConversationScreen(
             }
         },
         onBackButtonClick = { conversationScreenOnBackButtonClick(messageComposerViewModel, focusManager, navigator) },
-        composerMessages = messageComposerViewModel.infoMessage,
+        composerMessages = sendMessageViewModel.infoMessage,
         conversationMessages = conversationMessagesViewModel.infoMessage,
         conversationMessagesViewModel = conversationMessagesViewModel,
         onSelfDeletingMessageRead = messageComposerViewModel::startSelfDeletion,
         onNewSelfDeletingMessagesStatus = messageComposerViewModel::updateSelfDeletingMessages,
         tempWritableImageUri = messageComposerViewModel.tempWritableImageUri,
         tempWritableVideoUri = messageComposerViewModel.tempWritableVideoUri,
-        onFailedMessageRetryClicked = messageComposerViewModel::retrySendingMessage,
+        onFailedMessageRetryClicked = sendMessageViewModel::retrySendingMessage,
         requestMentions = messageComposerViewModel::searchMembersToMention,
         onClearMentionSearchResult = messageComposerViewModel::clearMentionSearchResult,
         onPermissionPermanentlyDenied = {
@@ -470,8 +473,8 @@ fun ConversationScreen(
         }
     )
     AssetTooLargeDialog(
-        dialogState = messageComposerViewModel.assetTooLargeDialogState,
-        hideDialog = messageComposerViewModel::hideAssetTooLargeError
+        dialogState = sendMessageViewModel.assetTooLargeDialogState,
+        hideDialog = sendMessageViewModel::hideAssetTooLargeError
     )
     VisitLinkDialog(
         dialogState = messageComposerViewModel.visitLinkDialogState,
@@ -489,15 +492,15 @@ fun ConversationScreen(
     )
 
     SureAboutMessagingInDegradedConversationDialog(
-        dialogState = messageComposerViewModel.sureAboutMessagingDialogState,
-        sendAnyway = messageComposerViewModel::acceptSureAboutSendingMessage,
-        hideDialog = messageComposerViewModel::dismissSureAboutSendingMessage
+        dialogState = sendMessageViewModel.sureAboutMessagingDialogState,
+        sendAnyway = sendMessageViewModel::acceptSureAboutSendingMessage,
+        hideDialog = sendMessageViewModel::dismissSureAboutSendingMessage
     )
 
-    (messageComposerViewModel.sureAboutMessagingDialogState as? SureAboutMessagingDialogState.Visible.ConversationUnderLegalHold)?.let {
+    (sendMessageViewModel.sureAboutMessagingDialogState as? SureAboutMessagingDialogState.Visible.ConversationUnderLegalHold)?.let {
         LegalHoldSubjectMessageDialog(
-            dialogDismissed = messageComposerViewModel::dismissSureAboutSendingMessage,
-            sendAnywayClicked = messageComposerViewModel::acceptSureAboutSendingMessage,
+            dialogDismissed = sendMessageViewModel::dismissSureAboutSendingMessage,
+            sendAnywayClicked = sendMessageViewModel::acceptSureAboutSendingMessage,
         )
     }
